requirements.txt
flask
sqlalchemy
psycopg2-binary 
playwright
pandas
openpyxl
pytest
requests
pillow
aiohttp
asyncpg